move HasMedia trait to traits folder
